feat: extract ingest routes to routes/ingest.py
